Refactoring modules some more
